requirements.txt
azure-functions
azure-identity
azure-storage-blob
azure-keyvault-secrets
azure-ai-projects
azure-data-tables
requests
exchangelib
